consume messages and register temperature value on database

+++ src/domain/TemperatureRepository.go
@@ -14,7 +14,7 @@ type TemperatureRepositoryDTO struct {
 	ID                 string    `json:"_id" bson:"_id"`
 	SensorID           string    `json:"sensor_id" bson:"sensor_id"`
 	CreatedAt          time.Time `json:"created_at" bson:"created_at"`
-	Value              int64     `json:"value"`
+	Value              float64   `json:"value"`
 	Percentage         float64   `json:"percentage"`
 	ExponentialAverage []float64 `json:"exponential_average"`
 	MovelAverage       []float64 `json:"movel_average"`
